Organizing shortcuts and CM commands
